+++ src/database/queries/proposals/checkExistsById.sql
@@ -1,0 +1,3 @@
+SELECT EXISTS (
+  SELECT 1 FROM proposals where id = :id
+) AS result