+++ src/queryParameters/extractSearchParameters.ts
@@ -4,7 +4,7 @@ import type { Request } from 'express';
 export const extractSearchParameters = (request: Request) => ({
 	/* eslint-disable no-underscore-dangle */
 	search:
-		request.query._content ??
+		request.query._content?.toString() ??
 		apiSpecification.components.parameters.searchParam.schema.default,
 	/* eslint-enable no-underscore-dangle */
 });