+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.5.0 - 2024-04-27
+
+### Added
+
+- The `OrganizationProposal` entity now exists.
+- The `OrganizationProposalBundle` entity now exists.
+- It is now possible to filter `/proposals` in terms of organization.
+- It is now possible to filter `/organizations` in terms of proposal.
+
 ## 0.4.0 - 2024-02-29
 
 ### Removed